Close the Height/MIN_HISTORY signed-unsigned class + operator[] hazards
The 1c07594 fix addressed a single signed/unsigned comparison and three
operator[] sites in the LoadBlockIndexDB / InitBlockIndex path. The same
bug class lived at eight more sites across main.cpp, triesync.cpp and
purge.cpp and would have crashed the daemon on a fresh or very young
chain the next time it was triggered.

* main.cpp 1693, 1727 (UpdateMissing / GetTotalMissing): 'Height() -
  MIN_HISTORY' promoted -1 to a huge unsigned value on an empty chain.
  Compute in int64_t so the empty case falls into the < 1 guard.
* main.cpp 2289 (ProcessBlockHeader): MIN_HISTORY was cast but
  chainActive.Height() was not, leaving the comparison signed/unsigned
  on an empty chain. Cast both sides.
* triesync.cpp 317-320 (TrieSync::GetSlice): the slice-anchor block
  chainHeaders[Height() - MIN_HISTORY] would have indexed into vChain
  with a huge unsigned index, dereferencing vChain far past its end --
  a guaranteed SIGSEGV the first time a peer asked for a slice on a
  fresh datadir. Guard explicitly, return an empty slice on an empty /
  very young chain.
* triesync.cpp 74 (TrieSync::Update): loop bound mixed int and uint64_t;
  cast MIN_HISTORY so the promotion stays in the signed domain.
* purge.cpp 52, 63, 73: chainActive.Height() and pindex->nHeight
  compared against MIN_HISTORY without a cast on either side.
* main.cpp 1336 (ConnectBlock): mapBlockIndex[block.hashPrevBlock] was
  the same operator[]-inserts-null pattern as the genesis fix; convert
  to a non-inserting lookup and DoS-score the peer on miss.
* main.cpp 2918-2921 (LoadBlockIndexDB): the deferred-path summary now
  emits 'trieOnline=... hashBestChain=<deferred> height=-1' so operators
  can grep the log for a clean cold-start signal.
* main.cpp 3075-3086 (InitBlockIndex): AcceptBlock(genesis) failure is
  now logged at WARN with the reject reason instead of being silently
  dropped; the pviewTip->Apply call below remains the actual fail-stop.
* main.cpp 2831, 2870, 3080, 3087: use the existing LookupBlockIndex
  helper instead of the open-coded find-then-ternary blocks.
* main.cpp 127-131: document the three ownership sites for
  pindexGenesisBlock.
* main.cpp 2875-2879, 2908-2913: document the syncPoint -> accelerated-
  startup SetTip ordering so the two SetTip calls don't read as a fight.
* init.cpp 858-913: collapse the dead do{ }while(false) retry loop
  (the noui prompt it gated on was a no-op) into a single load with a
  FailLoad lambda owning the failure log + InitError. -64 lines.

Verified:
  * make -j: clean (no relevant warnings).
  * src/test/test_cryptonite: 76/76 pass, no errors.
  * Wiped-datadir cold start: genesis accepted, trie.dat seeded
    (101 bytes), LoadBlockIndexDB summary on deferred path visible,
    AcceptBlock failure WARNING logged, clean exit (no SIGABRT, no
    SIGSEGV, no 'Incorrect or no genesis block found').
  * Restart: persisted trie loaded (878 bytes), trieOnline=1, success-
    path summary visible, clean exit.
  * Same daemon run reached height 1277 before kill, no abort.

+++ src/purge.cpp
@@ -49,7 +49,13 @@ void PurgeDB(){
     }
 
     printf("Tip is block %ld\n", chainActive.Height());
-    if(chainActive.Height() <= MIN_HISTORY){
+    // chainActive.Height() returns -1 on an empty chain; MIN_HISTORY is
+    // uint64_t. Promote both sides to int64_t so the comparison stays in
+    // the signed domain and the empty-chain case falls into the "not
+    // enough history" branch instead of evaluating as -1 <= 1000 (which
+    // implicit promotion renders TRUE anyway, but the explicit cast
+    // documents intent and survives any future type change).
+    if((int64_t)chainActive.Height() <= (int64_t)MIN_HISTORY){
 	printf("Not enough history to prune\n");
 	goto fail;
     }
@@ -59,8 +65,11 @@ void PurgeDB(){
 
     //mapBlockIndex contains all blocks, so we can read those for info
     for (std::pair<uint256, CBlockIndex*> item : mapBlockIndex){
-	CBlockIndex *pindex = item.second;	
-	if(pindex->nHeight + MIN_HISTORY >= chainActive.Height()){
+	CBlockIndex *pindex = item.second;
+	// MIN_HISTORY is uint64_t; promote both sides to int64_t to keep the
+	// comparison in the signed domain (see chainActive.Height() comment
+	// above).
+	if((int64_t)pindex->nHeight + (int64_t)MIN_HISTORY >= (int64_t)chainActive.Height()){
 	    //Block cannot be deleted
 	    if (pindex->nStatus & BLOCK_HAVE_DATA || pindex->nStatus & BLOCK_HAVE_UNDO)
 		setNeed.insert(pindex->nFile);
@@ -69,8 +78,8 @@ void PurgeDB(){
 
     //Second pass, locate all blocks that can have transactions deleted from txindex
     for (std::pair<uint256, CBlockIndex*> item : mapBlockIndex){
-	CBlockIndex *pindex = item.second;	
-	if(pindex->nHeight + MIN_HISTORY < chainActive.Height()){
+	CBlockIndex *pindex = item.second;
+	if((int64_t)pindex->nHeight + (int64_t)MIN_HISTORY < (int64_t)chainActive.Height()){
 	    //Block can be deleted
 	    if (pindex->nStatus & BLOCK_HAVE_DATA || pindex->nStatus & BLOCK_HAVE_UNDO)
 		//Block has TX's
